Makefile: reorganize
Reorganize the main Makefile in order to:
1. be able to build in parallel
2. have build failures in subdirectories stop the build

Now the "-j <#threads>" option can be used with 'make'.

Also, if a build failure occurs in a subdirectory, this new arrangement will
cause the build to stop and fail instead of silently ignoring it.

To build simply invoke 'make' with or without a -j option.
To install simply invoke: make TARGET=install
To clean simply invoke: make TARGET=clean

Signed-off-by: Trevor Woerner <twoerner@gmail.com>